\name{axisPhylo2}
\alias{axisPhylo2}
\title{axisPhylo with more flexibility in labeling}
\usage{
  axisPhylo2(side = 1, roundlabels = FALSE, minage = 0,
    ...)
}
\arguments{
  \item{side}{The side to plot on (default 1, bottom)}

  \item{roundlabels}{Number of digits to round to, if
  desired}

  \item{minage}{Starting age, if desired}

  \item{...}{Additional arguments to standard functions}
}
\value{
  nothing
}
\description{
  Hacking axisPhylo to make it more flexible
}
\note{
  Go BEARS!
}
\examples{
testval=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}
  \url{https://code.google.com/p/lagrange/}

  FosterIdiots
}

