\name{BioGeoBEARS_model_object_to_init_params}
\alias{BioGeoBEARS_model_object_to_init_params}
\title{Produce initial parameters from a BioGeoBEARS model object}
\usage{
  BioGeoBEARS_model_object_to_init_params(BioGeoBEARS_model_object)
}
\arguments{
  \item{BioGeoBEARS_model_object}{The BioGeoBEARS_model
  object, of class \code{BioGeoBEARS_model}}
}
\value{
  \code{params} parameter vector
}
\description{
  What it says.
}
\note{
  Go BEARS!
}
\examples{
test=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link[BioGeoBEARS]{define_BioGeoBEARS_model_object}}
}

