\name{BioGeoBEARS_run}
\alias{BioGeoBEARS_run}
\title{An object of class BioGeoBEARS_run holding the model inputs}
\description{
  An object of class BioGeoBEARS_run holding the model
  inputs
}
\note{
  Go BEARS!
}
\section{Slots}{
  \describe{ \item{\code{list}:}{List of class
  \code{"list"}, containing inputs list from
  define_BioGeoBEARS_run} }
}
\examples{
test=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link{define_tipranges_object}},
  \code{\link{getareas_from_tipranges_object}},
  \code{\link[cladoRcpp]{areas_list_to_states_list_old}},
  \code{\link{areas_list_to_states_list_new}},
  \code{\link{tipranges_to_tip_condlikes_of_data_on_each_state}}
}

