\name{corner_coords}
\alias{corner_coords}
\title{Get the corner coordinates}
\usage{
  corner_coords(tr)
}
\arguments{
  \item{tr}{A tree object in \code{\link[ape]{phylo}}
  format.}
}
\value{
  \code{corners_list}
}
\description{
  Gets the coordinates of the corners when the tree is
  plotted.
}
\note{
  Go BEARS!
}
\examples{
blah=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link[ape]{phylo}}, \code{\link{get_nodenums}}
}

