\name{default_states_list}
\alias{default_states_list}
\title{Default input for a states_list}
\usage{
  default_states_list()
}
\value{
  \code{states_list} The list of states
}
\description{
  R CMD check limits the length of inputs to variables for
  functions; this is a workaround.
}
\note{
  Go BEARS!
}
\examples{
states_list = default_states_list()
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link{make_dispersal_multiplier_matrix}}
}

