\name{get_APE_nodenums}
\alias{get_APE_nodenums}
\title{Get R internal node numbers}
\usage{
  get_APE_nodenums(tr)
}
\arguments{
  \item{tr}{A \code{\link[ape]{phylo}} tree object}
}
\value{
  \code{nodenums} A list of node numbers
}
\description{
  Utility function
}
\note{
  Go BEARS!
}
\examples{
test=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link{get_lagrange_nodenums}}, \code{\link{prt}}
}

