\name{get_colors_for_numareas}
\alias{get_colors_for_numareas}
\title{Get colors for a certain number of single areas}
\usage{
  get_colors_for_numareas(numareas, use_rainbow = FALSE)
}
\arguments{
  \item{numareas}{The number of areas}

  \item{use_rainbow}{If TRUE, force use of
  \code{rainbow()}}
}
\value{
  \code{colors_matrix} The colors for the single areas, 1
  column per area
}
\description{
  Like it says.
}
\note{
  Go BEARS!
}
\examples{
testval=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link[stats]{optim}}
}

