\name{get_node_ages_of_tips}
\alias{get_node_ages_of_tips}
\title{Get the ages of each tip above the root}
\usage{
  get_node_ages_of_tips(obj)
}
\arguments{
  \item{obj}{An ape phylo object}
}
\value{
  \code{TF_tips} The age (from the root) of each tip.
}
\description{
  A utility function.
}
\note{
  Go BEARS!
}
\examples{
test=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link{prt}}, \code{\link{chainsaw2}}
}

