\name{get_nodenums}
\alias{get_nodenums}
\title{Get the unique node numbers in a tree}
\usage{
  get_nodenums(t)
}
\arguments{
  \item{t}{A tree object in \code{\link[ape]{phylo}}
  format.}
}
\value{
  \code{ordered_nodenames} The node numbers, in order.
}
\description{
  This is a utility function for
  \code{\link{get_nodenum_structural_root}}.
}
\note{
  Go BEARS!
}
\examples{
blah = 1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link[ape]{phylo}},
  \code{\link{get_nodenum_structural_root}}
}

