\name{get_parent}
\alias{get_parent}
\title{Get the direct parent node of a node}
\usage{
  get_parent(nodenum, t)
}
\arguments{
  \item{nodenum}{The node number to get the parent of}

  \item{t}{An ape phylo object}
}
\value{
  \code{parent_nodenum}The parent node number
}
\description{
  Get the direct parent node of a node
}
\note{
  Go BEARS!
}
\examples{
test=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link{findall}}, \code{\link{chainsaw2}}
}

