\name{get_TF_tips}
\alias{get_TF_tips}
\title{Get TRUE/FALSE for nodes being tips}
\usage{
  get_TF_tips(obj)
}
\arguments{
  \item{obj}{An ape phylo object}
}
\value{
  \code{TF_tips} The TRUE/FALSE list for each tip.
}
\description{
  A utility function
}
\note{
  Go BEARS!
}
\examples{
test=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link{prt}}, \code{\link{chainsaw2}}
}

