\name{is.not.na}
\alias{is.not.na}
\title{Check for not NA}
\usage{
  is.not.na(x)
}
\arguments{
  \item{x}{Thing to check for NA}
}
\value{
  \code{TRUE} or \code{FALSE}
}
\description{
  A utility function.
}
\note{
  Go BEARS!
}
\examples{
test=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link{prt}}, \code{\link{chainsaw2}}
}

