\name{label_nodes_postorder_phylo3}
\alias{label_nodes_postorder_phylo3}
\title{Add postorder node number labels to a phylo3 tree object.}
\usage{
  label_nodes_postorder_phylo3(tr2)
}
\arguments{
  \item{tr2}{\code{\link[ape]{phylo}} tree object.}
}
\value{
  \code{tr2} A \code{\link[ape]{phylo}} tree object with
  node labels added.
}
\description{
  Adds \code{\link[phylobase]{phylobase}}
  \code{\link[phylobase]{phylo4}} postorder node number
  labels to a \code{\link[ape]{phylo}} tree object.
}
\note{
  Go BEARS!
}
\examples{
test=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link[ape]{phylo}}, \code{\link[phylobase]{phylo4}}
}

