\name{LapDem_Model_calc_logPP_orig}
\alias{LapDem_Model_calc_logPP_orig}
\title{The function calculating posterior probability of parameter values}
\usage{
  LapDem_Model_calc_logPP_orig(params, MyData)
}
\arguments{
  \item{params}{the parameter values}

  \item{MyData}{The
  \code{\link[LaplacesDemon]{LaplacesDemon}} input data.}
}
\value{
  \code{Modelout} The MCMC output.
}
\description{
  What it says
}
\note{
  Go BEARS!
}
\examples{
test=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  LaplacesDemon_Tutorial

  Matzke_2012_IBS
}
\seealso{
  \code{\link[LaplacesDemon]{LaplacesDemon}}
}

