\name{list2str}
\alias{list2str}
\title{Convert a list of items to a string}
\usage{
  list2str(list1, spacer = " ")
}
\arguments{
  \item{list1}{The list to convert.}

  \item{spacer}{The space between each item. Default " ".}
}
\value{
  \code{tmpstr} The output string.
}
\description{
  This is a shortcut to save time when converting a list of
  items to a string.
}
\note{
  Go BEARS!
}
\examples{
test=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}
  \url{https://code.google.com/p/lagrange/}

  Matzke_2012_IBS

  ReeSmith2008

  FosterIdiots
}
\seealso{
  \code{\link[base]{paste}},
  \code{\link[base]{as.character}}
}

