\name{merge_words_nonwords}
\alias{merge_words_nonwords}
\title{Merge lists of words and nonwords (numbers) that may be of different length}
\usage{
  merge_words_nonwords(words, nonwords)
}
\arguments{
  \item{words}{A list of words}

  \item{nonwords}{A list of nonwords}
}
\value{
  \code{sentence} A text string.
}
\description{
  Utility function.
}
\note{
  Go BEARS!
}
\examples{
test=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link[BioGeoBEARS]{define_BioGeoBEARS_model_object}}
}

