\name{order_tipranges_by_tr}
\alias{order_tipranges_by_tr}
\title{Order the tipranges in a tipranges object so they match the order of tips in a tree}
\usage{
  order_tipranges_by_tr(tipranges, tr)
}
\arguments{
  \item{tipranges}{A tipranges object.}

  \item{tr}{An ape tree object.}
}
\value{
  \code{tipranges} The reordered data.frame
}
\description{
  Utility function. What it says.  Life can get very
  confusing if you don't do this before plotting.
}
\note{
  Go BEARS!
}
\examples{
test=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}
  \url{https://code.google.com/p/lagrange/}

  Matzke_2012_IBS
}
\seealso{
  \code{\link[base]{unlist}}
}

