\name{params_into_BioGeoBEARS_model_object}
\alias{params_into_BioGeoBEARS_model_object}
\title{Feed modified parameters back into a BioGeoBEARS model object}
\usage{
  params_into_BioGeoBEARS_model_object(BioGeoBEARS_model_object,
    params)
}
\arguments{
  \item{BioGeoBEARS_model_object}{The BioGeoBEARS_model
  object, of class \code{BioGeoBEARS_model}}

  \item{params}{parameter vector}
}
\value{
  \code{BioGeoBEARS_model_object} The BioGeoBEARS_model
  object, of class \code{BioGeoBEARS_model}
}
\description{
  What it says.
}
\note{
  Go BEARS!
}
\examples{
test=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link[BioGeoBEARS]{define_BioGeoBEARS_model_object}}
}

