\name{prt_tree_to_phylo4}
\alias{prt_tree_to_phylo4}
\title{prt_tree_to_phylo4}
\usage{
  prt_tree_to_phylo4(prt_tr)
}
\arguments{
  \item{prt_tr}{A prt_tree from the function
  \code{\link{prt}}.}
}
\value{
  \code{newtr} A \code{\link[phylobase]{phylobase}}
  \code{\link[phylobase]{phylo4}} tree object.
}
\description{
  Converts a tree table (a prt_tree from the function
  \code{\link{prt}}, which prints trees to tables) to a
  \code{\link[phylobase]{phylobase}}
  \code{\link[phylobase]{phylo4}} tree object.
}
\note{
  Go BEARS!
}
\examples{
test=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link[phylobase]{phylo4}}, \code{\link{prt}}
}

