\name{section_the_tree}
\alias{section_the_tree}
\title{Section a tree for stratified analysis}
\usage{
  section_the_tree(inputs)
}
\arguments{
  \item{inputs}{The list of inputs for stratified analysis}
}
\value{
  \code{inputs} with inputs$tree_sections_list added.
}
\description{
  A utility function for stratified analysis.
}
\note{
  Go BEARS!
}
\examples{
test=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link{prt}}, \code{\link{chainsaw2}}
}

