\name{calc_linked_params_BioGeoBEARS_model_object}
\alias{calc_linked_params_BioGeoBEARS_model_object}
\title{new_BioGeoBEARS_model_object = calc_linked_params_BioGeoBEARS_model_object(BioGeoBEARS_model_object)
new_BioGeoBEARS_model_object}
\usage{
  calc_linked_params_BioGeoBEARS_model_object(BioGeoBEARS_model_object)
}
\arguments{
  \item{BioGeoBEARS_model_object}{The BioGeoBEARS_model
  object, of class \code{BioGeoBEARS_model}}
}
\value{
  \code{BioGeoBEARS_model_object} (Updated version of) the
  BioGeoBEARS_model object, of class
  \code{BioGeoBEARS_model}
}
\description{
  Update parameters that are deterministic functions of
  free parameters
}
\details{
  What it says.
}
\note{
  Go BEARS!
}
\examples{
test=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link[BioGeoBEARS]{define_BioGeoBEARS_model_object}}
}

