\name{colors_legend}
\alias{colors_legend}
\title{Plot a colors legend for geographic ranges}
\usage{
  colors_legend(possible_ranges_list_txt,
    colors_list_for_states, ncol = 2, cex = 2.5)
}
\arguments{
  \item{possible_ranges_list_txt}{A list of the allowed
  ranges/states}

  \item{colors_list_for_states}{The corresponding colors}

  \item{ncol}{Number of columns in the legend, default 2}

  \item{cex}{Text size multiplier, default 2.5}
}
\value{
  Nothing
}
\description{
  Like it says.
}
\note{
  Go BEARS!
}
\examples{
testval=1
}
\author{
  Nicholas J. Matzke \email{matzke@berkeley.edu}
}
\references{
  \url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}

  Matzke_2012_IBS
}
\seealso{
  \code{\link[stats]{optim}}
}

